finetuning/requirements.txt
torch
pandas
apollo-torch
huggingface_hub
datasets
python-dotenv

